+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "basson"
-version = "0.1dev"
+version = "0.0.2"
 authors = [
    { name="Leo 52044", email="lev.portnyagin@vk.com" },]
 description = "Easy-to-use wrapper for BASS audio library"
